libstdc++ c++2a is now using structured bindings, and those don't
	work :(

From-SVN: r279041

+++ gcc/testsuite/gcc.dg/lto/pr88077_0.c
@@ -1,3 +1,3 @@
 /* { dg-lto-do link } */
 
-int HeaderStr;
+long HeaderStr;
